Extract orbit-store package from orbit-common.
