+++ build-support/globalize-orbit-store.js
@@ -1,0 +1,1 @@
+self.OrbitStore = requireModule("orbit-store");
