+++ src/orbit-store.js
@@ -1,0 +1,4 @@
+import Store from 'orbit-store/store';
+import Cache from 'orbit-store/cache';
+
+export { Store, Cache };
