+++ src/orbit-store/cache/inverse-transforms.js
@@ -1,6 +1,5 @@
-import { toIdentifier, parseIdentifier } from '../lib/identifiers';
 import { eq } from 'orbit/lib/eq';
-
+import { toIdentifier, parseIdentifier } from 'orbit-common/lib/identifiers';
 
 const InverseTransforms = {
   addRecord(doc, op) {
