+++ src/orbit-store/cache/observables/records-observable.js
@@ -5,7 +5,6 @@ import 'rxjs/add/operator/mergeMap';
 import 'rxjs/add/operator/concatAll';
 import 'orbit-common/rxjs/add/operator/matching';
 
-
 export default class RecordsObservable extends Observable {
   constructor(subscribe, cache) {
     super(subscribe);
