+++ test/tests/orbit-store/unit/cache/live-queries-test.js
@@ -1,4 +1,3 @@
-import Cache from 'orbit-common/cache';
 import Schema from 'orbit-common/schema';
 import KeyMap from 'orbit-common/key-map';
 import { identity } from 'orbit-common/lib/identifiers';
@@ -14,6 +13,7 @@ import {
   // replaceHasMany,
   replaceHasOne
 } from 'orbit-common/transform/operators';
+import Cache from 'orbit-store/cache';
 
 const planetsSchema = new Schema({
   models: {
